V2 (#208)
Internal API rewrite, no breaking changes.

Legacy code rewrite to simplify library maintenance.

graph_id and legend_toggle props were renamed to graphID and legendToggle for consistency with the others. But support for the old prop names will remain for the foreseeable future. 

loadCharts props was removed. The lib will check to see if google charts loader was  loaded in the DOM.